requirements.txt
APScheduler==3.6.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
cloudinary==1.17.0
coralogix-logger==2.0.1.post7
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.5
django-cloudinary-storage==0.2.3
django-cors==0.1
django-cors-headers==3.1.0
django-filter==2.2.0
django-guardian==2.1.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-sendgrid-v5==0.8.1
djangorestframework==3.10.3
emoji==0.5.3
enum34==1.1.6
future==0.18.2
gunicorn==20.0.4
idna==2.7
itypes==1.1.0
Jinja2==2.10.1
Markdown==3.1.1
MarkupSafe==1.1.1
mock==3.0.5
msgpack==1.0.0
openapi-codec==1.3.2
pdfkit==0.6.1
Pillow==7.1.1
psycopg2-binary==2.8.5
Pygments==2.4.2
PyJWT==1.7.1
PySocks==1.7.0
python-http-client==3.2.7
python-magic==0.4.15
pytz==2019.2
redis==3.3.8
requests==2.20.0
rq==1.1.0
sendgrid==6.2.2
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
stripe==2.35.1
timber==2.1.0
twilio==6.30.0
tzlocal==2.0.0
uritemplate==3.0.0
urllib3==1.24.3
zipfile37==0.1.3
